orders, order items, product qty update functional
Database is working for persisting customer orders, order items and is
also updating products qty to reflect what was purchased.

We have to create a confirmation and receipt page that queries the
database for items ordered and total amount of order.

After that we have to create a link to allow a logged in user to view
their orders and create an admin link for an admin to view all orders
by all customers.